+++ src/utils/index.js
@@ -1,6 +1,8 @@
 import { unique } from '@flekschas/utils';
 import * as PIXI from 'pixi.js';
 
+export { default as scaleLinear } from './scale-linear';
+
 export const cloneSprite = sprite => {
   const clonedSprite = new PIXI.Sprite(sprite.texture);
   clonedSprite.interactive = sprite.interactive;
@@ -73,53 +75,6 @@ export const ifNotNull = (v, alternative = null) =>
  * @return  {array}  Quadruple representing the homogeneous vector
  */
 export const toHomogeneous = (x, y) => [x, y, 0, 1];
-
-export const scaleLinear = () => {
-  let domainMin = 0;
-  let domainMax = 1;
-  let domainSize = 1;
-
-  let rangeMin = 0;
-  let rangeMax = 1;
-  let rangeSize = 1;
-
-  const scale = value =>
-    Math.min(
-      rangeMax,
-      Math.max(
-        rangeMin,
-        rangeMax - ((domainMax - value) / domainSize) * rangeSize
-      )
-    );
-
-  scale.domain = (newDomain = []) => {
-    if (newDomain.length === 0) return [domainMin, domainMax];
-
-    const [newDomainMin, newDomainMax] = newDomain;
-
-    domainMin = newDomainMin;
-    domainMax = newDomainMax;
-
-    domainSize = domainMax - domainMin || 1;
-
-    return scale;
-  };
-
-  scale.range = (newRange = []) => {
-    if (newRange.length === 0) return [rangeMin, rangeMax];
-
-    const [newRangeMin, newRangeMax] = newRange;
-
-    rangeMin = newRangeMin;
-    rangeMax = newRangeMax;
-
-    rangeSize = rangeMax - rangeMin;
-
-    return scale;
-  };
-
-  return scale;
-};
 
 export const colorToDecAlpha = (color, defaultAlpha = 1) => {
   if (typeof color === 'string' || color instanceof String) {
